+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.28](https://github.com/medly/medly-components/compare/@medly-components/icons@1.0.0-beta.27...@medly-components/icons@1.0.0-beta.28) (2020-06-05)
+
+**Note:** Version bump only for package @medly-components/icons
+
+
+
+
+
 # [1.0.0-beta.27](https://github.com/medly/medly-components/compare/@medly-components/icons@1.0.0-beta.26...@medly-components/icons@1.0.0-beta.27) (2020-06-01)